API/Reference

ADI.reconstruct — Function

reconstruct(::ADIAlgorithm, cube; [ref], kwargs...)

Reconstruct the PSF approximation for the given algorithm, using ref as the reference cube if given and supported by the algorithm.

Examples

julia> cube, angles = # load data

julia> S = reconstruct(PCA(10), cube);

julia> size(S) == size(cube)
true

julia> flat_res = collapse(cube .- S, angles); # form resid, derotate, and combine

ADI.subtract — Function

subtract(::ADIAlgorithm, cube; [ref], kwargs...)

Reconstruct the PSF approximation for the given algorithm and subtract it from cube, using ref as the reference cube if given and supported by the algorithm.

Examples

julia> cube, angles = # load data

julia> R = subtract(PCA(10), cube);

julia> size(R) == size(cube)
true

julia> flat_res = collapse(R, angles); # derotate, and combine

ADI.fit — Function

ADI.fit(::ADIAlgorithm, cube; [ref], kwargs...)

Given the description of an algorithm and the appropriate options, take the pixels from cube and fit them, returning an (ADIDesign) containing the necessary information from the fit (e.g. the principal components from PCA decomposition).

If the algorithm supports reference differential imaging (RDI), the reference cube can be passed by the keyword argument ref.

ADI.design — Function

ADI.design(::ADIDesign)

Return the pertinent data required to form the PSF approximation. For example, weights and components for PCA/NMF or the median frame for classic ADI.

ADI.ADIDesign — Type

ADI.ADIDesign

An abstract type used for holding the output of ADI.fit. The purpose of these types is to hold the minimum information required to reconstruct the PSF estimate (e.g. weights and principal components from PCA). ADI.jl includes two designs- ADI.ClassicDesign and ADI.LinearDesign.

Interface

When implementing a new algorithm, if your outputs do not fit into either of those designs, you will have to create your own ADIDesign with the following methods-

ADI.design(::Design)

accessor for the pertinent data to approximate the PSF

reconstruct(::Design)

return the approximate PSF estimate as a matrix

ADI.LinearDesign — Type

ADI.LinearDesign(basis, coeffs)

A "linear" design implies the use of some linear basis for reconstructing data along with a set of coefficients or weights. The reconstruction will be the matrix product of the weights and the basis, w * Z.

ADI.design will return (basis, ceoffs), and you can also extract them via iteration, like Z, w = design.