Classic

API/Reference

ADI.Classic — Type

Classic(;method=median)
Classic(method)

Classic PSF subtraction using the median of entire data cube. If another statistic is desired, like the mean, it can be passed as an argument as long as it supports slicing multi-dimensional arrays.

References

Marois et al. 2006 Angular Differential Imaging: A Powerful High-Contrast Imaging Technique

ADI.ClassicDesign — Type

ADI.ClassicDesign(n, frame)

Output for the Classic algorithm which contains the static frame unrolled into a vector (with size (1, Npx)). reconstruct will tile this vector in a non-allocating way n times to appear like a flattened cube. ADI.design will return the static frame .