GreeDS

API/Reference

ADI.GreeDS — Type

GreeDS(alg=PCA(); threshold=0.0, progress=true)
GreeDS(ncomps; threshold=0.0, progress=true)

Performs the greedy disk subtraction (GreeDS) algorithm.

This method is an iterative approach to standard ADI reduction which seeks to minimize over-subtraction by constraining the low-rank matrix approximation from alg.

For large data cubes the iteration can cause slowdowns, so a progress bar is provided using the ProgressLogging API along with the progress keyword. It won't appear without a logging backend, such as TerminalLoggers.

Algorithms

Although the original paper explicitly uses PCA, we allow use of any linear ADI algorithm that is characterized by ncomps. By default, uses PCA. If ncomps are provided, will default to PCA as well.

References

Pairet et al. 2018 "Reference-less algorithm for circumstellar disks imaging"

Pairet et al. 2020 "MAYONNAISE: a morphological components analysis pipeline for circumstellar disks and exoplanets imaging in the near infrared"