LOCI

API/Reference

ADI.LOCI — Type

LOCI(; dist_threshold=nothing, metric=Cityblock())

Local optimal combination of images (LOCI).

If provided, the frames used for the reference are filtered to only include the frames whose pairwise distances are within the dist_threshold quantile. The distances are measured using Distances.jl and the metric used for measuring the distacnes can be specified with metric (by default uses the Manhattan/cityblock metric).

Traditional LOCI

Unless LOCI is applied Framewise, no distance thresholding will occur. In order to recreate the traditional LOCI algorithm, consider constructing an algorithm like

# only use the most-similar 90th-percentile frames
alg = Framewise(LOCI(dist_threshold=0.90))

References

Lafreniere et al. (2007) "A New Algorithm for Point-Spread Function Subtraction in High-Contrast Imaging: A Demonstration with Angular Differential Imaging"