API/Reference

For now, here is a dump of all documented functions and types.

Index

ADI.ADIAlgorithm

ADI.ADIAlgorithm

ADI.GreeDS

ADI.LinearAlgorithm

ADI.Median

ADI.PCA

ADI.decompose

ADI.pairet_theta

ADI.reconstruct

API/Reference

ADI.GreeDS — Type

GreeDS(alg=PCA(); threshold=0.0, progress=true)

Performs the greedy disk subtraction (GreeDS) algorithm.

This method is an iterative approach to standard ADI reduction which seeks to minimize over-subtraction by constraining the low-rank matrix approximation from alg.

For large data cubes the iteration can cause slowdowns, so a progress bar is provided using the ProgressLogging API along with the progress keyword. It won't appear without a logging backend, such as TerminalLoggers.

Algorithms

Although the original paper explicitly uses PCA, we allow use of any linear ADI algorithm that is characterized by ncomps. By default, uses PCA.

References

Pairet et al. 2018 "Reference-less algorithm for circumstellar disks imaging"

Pairet et al. 2020 "MAYONNAISE: a morphological components analysis pipeline for circumstellar disks and exoplanets imaging in the near infrared"

ADI.Median — Type

Median()

Classic PSF subtraction using the median of entire data cube.

References

Marois et al. 2006 Angular Differential Imaging: A Powerful High-Contrast Imaging Technique

ADI.PCA — Type

PCA(;ncomps=nothing, pratio=1)

If ncomps is nothing, it will be set to the number of frames in the reference cube when processed.

References

Soummer, Pueyo, and Larkin (2012) "Detection and Characterization of Exoplanets and Disks Using Projections on Karhunen-Loève Eigenimages"

ADI.decompose — Function

ADI.decompose(::LinearAlgorithm, cube, angles, [cube_ref]; kwargs...)

ADI.reconstruct — Function

reconstruct(::ADIAlgorithm, cube, angles, [cube_ref]; kwargs...)

Reconstruct the PSF approximation for the given algorithm, using cube_ref as the reference cube if given.

ADI.ADIAlgorithm — Type

ADI.ADIAlgorithm <: Function

This abstract type is used for defining ADI algorithms. See the extended help (??ADIAlgorithm) for interface details.

Extended help

Interface

To extend ADIAlgorithm you may implement the following | function | default | description | |–––––|––––-|––––––-| | ADI.reconstruct | | Subroutine for creating the full reconstructed cube with the PSF | | (::ADIAlgorithm) | subtracts output of reconstruct, then derotates and collapses | Subroutine for returning the reduced residual cube |

ADI.ADIAlgorithm — Method

(::ADIAlgorithm)(cube, angles, [cube_ref]; kwargs...)

Fully process an ADI data cube using reconstruct and collapsing the residuals. Keyword arguments will be passed to HCIToolbox.collapse!.

ADI.LinearAlgorithm — Type

ADI.LinearAlgorithm <: ADI.ADIAlgorithm

This abstract type is used for defining linear ADI algorithms. See the extended help (??LinearAlgorithm) for interface details.

Extended help

Interface

To extend LinearAlgorithm you may implement the following

function	default	description

ADI.fit		Subroutine for fitting the linear basis and coefficients as unrolled matrices

ADI.reconstruct	Computes the inner product of the design matrix and weights from decompose	Subroutine for creating the full reconstructed cube with the PSF

(::LinearAlgorithm)	subtracts output of reconstruct, then derotates and collapses	Subroutine for returning the reduced residual cube

ADI.pairet_theta — Method

pairet_theta(frame, angles, threshold; kwargs...)

Takes a frame, expands it into a cube, rotates it clockwise by angles, and min-clips at threshold. Keyword arguments will be passed to HCIToolbox.derotate!.