GreeDS

API/Reference

ADI.GreeDS — Type

GreeDS(alg=PCA(); threshold=0)
GreeDS(ncomps; threshold=0, options...)

Performs the greedy disk subtraction (GreeDS) algorithm.

This method is an iterative approach to standard ADI reduction which seeks to minimize over-subtraction by constraining the low-rank matrix approximation from alg. By default, uses PCA. If ncomps or other PCA options are provided, they will be passed to the constructor.

Note

The GreeDS algorithm requires fully reconstructing a cube at each iteration, which requires knowing the geometry of the input (full-frame, annulus, etc.) and the corresponding parallactic angles. These angles must be passed as a keyword argument angles. In the case of reducing data, e.g. GreeDS()(cube, angles) the angles will be passed automatically. It is important to clarify, these angles should correspond to the reference data in the case of RDI, e.g. GreeDS()(cube, angles; ref=ref_cube, angles=ref_angles)

Algorithms

The following algorithms work natively with GreeDS: PCA, TPCA, and NMF

References

Pairet et al. 2018 "Reference-less algorithm for circumstellar disks imaging"

Pairet et al. 2020 "MAYONNAISE: a morphological components analysis pipeline for circumstellar disks and exoplanets imaging in the near infrared"