PCA

API/Reference

ADI.PCA — Type

PCA(;ncomps=nothing, pratio=1) <: LinearAlgorithm

Use principal components analysis (PCA) to form a low-rank orthonormal basis of the input. Uses deterministic singular-value decomposition (SVD) to decompose data.

If ncomps is nothing, it will be set to the number of frames in the reference cube when processed.

References

Soummer, Pueyo, and Larkin (2012) "Detection and Characterization of Exoplanets and Disks Using Projections on Karhunen-Loève Eigenimages"

Implements

decompose

ADI.TPCA — Type

TPCA(;ncomps=nothing) <: LinearAlgorithm

Perform principal components analysis (PCA) using truncated SVD (TSVD; provided by TSVD.jl) instead of deterministic SVD. This is often faster thant PCA but is non-determinstic. We find the differences unnoticable in practice.

If ncomps is nothing, it will be set to the number of frames in the reference cube when processed.

Implements

decompose

See Also

PCA, TSVD.tsvd