API/Reference

For now, here is a dump of all documented functions and types.

Index

ADI.PCADesign

ADI.pairet

ADI.pca

ADI.reconstruct

Base.reduce

API/Reference

ADI.pairet — Method

pairet([alg=pca], cube, angles; ncomps, threshold=0, kwargs...)

Performs the bootstrapping algorithm defined in Pairet et al. 2018.

This method is an iterative approach to standard ADI reduction which seeks to minimze over-subtracting signal in the low-rank approximation of cube.

ncomps can be an integer, which will iterate over 1:ncomps, otherwise it can be any sub-type of AbstractRange{<:Int}. As part of the Pairet algorithm, the low-rank approximation at each iteration will be min-clipped at threshold.

Any extra keyword arguments will be passed to reduce(::ADIDesign).

Algorithms

Although the original paper explicitly uses PCA, we allow use of any ADI algorithm that is characterized by ncomps. By default, uses pca.

pca

nmf (not yet implemented)

References

Pairet et al. 2018 "Reference-less algorithm for circumstellar disks imaging"

ADI.pca — Method

pca(cube, [ref], angles; ncomps, pratio=1)::PCADesign

Decomposes a cube (or reference) using principal component analysis (PCA) to form the approximate reconstruction of the systematic noise.

This will create a design matrix (the principal subspace) of the cube (or reference) truncated to either ncomps or until the prinicpal ratio is equal to pratio (whichever is fewer). As ncomps (or pratio) increase, more structure is removed from the cube, thus it is possible to over-subtract signal when choosing the size of the principal subpspace.

ADI.reconstruct — Method

reconstruct(::ADIDesign, cube) -> cube
reconstruct(::ADIDesign, matrix) -> matrix

Reconstrucst an approximation of the input using the design.

See Also

reduce

ADI.PCADesign — Type

PCADesign(A, w, reconstructed, S, angles, pratio)

A container for PCA-based ADI algorithm output.

A is the principal subspace, w are the weights used to reconstruct to the target cube. The residual from reconstruction is stored in S as a cube. The parallactic-angles are stored in angles. Finally, the principal ratio (the ratio of explained variance per component to the total explained variance) is stored in pratio.

See Alse

pca, pairet

Base.reduce — Method

reduce(::ADIDesign; method=median, deweight=true, fill=0) -> matrix
reduce(::ADIDesign, cube, [angles]; method=median, deweight=true, fill=0) -> matrix

Reduces an ADI Design matrix by computing the residual of the reconstructed cube and the target cube, then collapsing it. The keyword arguments will be passed to HCIToolbox.collapse!.

If cube and angles are provided, referential differential imaging (RDI) will be done by reconstructing cube using the input design. If angles are not provided, the same angles used for the design construction will be used.

See Also

reconstruct