API/Reference

For now, here is a dump of all documented functions and types.

Index

ADI.Metrics

ADI.ADIAlgorithm

ADI.DoubleSDI

ADI.GreeDS

ADI.LinearAlgorithm

ADI.Median

ADI.NMF

ADI.PCA

ADI.SDIAlgorithm

ADI.SingleSDI

ADI.TPCA

ADI.Metrics.contrast_curve

ADI.Metrics.detectionmap

ADI.Metrics.significance

ADI.Metrics.snr

ADI.Metrics.throughput

ADI.decompose

ADI.reconstruct

API/Reference

ADI.ADIAlgorithm — Type

ADI.ADIAlgorithm <: Function

This abstract type is used for defining ADI algorithms. See the extended help (??ADIAlgorithm) for interface details.

Extended help

Interface

To extend ADIAlgorithm you may implement the following

function	default	description

reconstruct		Subroutine for creating the full reconstructed cube with the PSF

(::ADIAlgorithm)	subtracts output of reconstruct, then derotates and collapses	Subroutine for returning the reduced residual cube

ADI.LinearAlgorithm — Type

ADI.LinearAlgorithm <: ADI.ADIAlgorithm

This abstract type is used for defining linear ADI algorithms. See the extended help (??LinearAlgorithm) for interface details.

Extended help

Interface

To extend LinearAlgorithm you may implement the following

function	default	description

decompose		Subroutine for fitting the linear basis and coefficients as unrolled matrices

reconstruct	Computes the inner product of the design matrix and weights from decompose	Subroutine for creating the full reconstructed cube with the PSF

(::LinearAlgorithm)	subtracts output of reconstruct, then derotates and collapses	Subroutine for returning the reduced residual cube

ADI.reconstruct — Function

reconstruct(::ADIAlgorithm, cube, angles, [cube_ref]; kwargs...)

Reconstruct the PSF approximation for the given algorithm, using cube_ref as the reference cube if given.

ADI.decompose — Function

ADI.decompose(::LinearAlgorithm, cube, angles, [cube_ref]; kwargs...)