Metrics

API/Reference

ADI.Metrics — Module

ADI.Metrics

This module provides code for analyzing the results from ADI in a way that is interpretable statistically. Some of the key functionalities are signal-to-noise, significance, the receiver operating characteristic, and the contrast curve.

ADI.Metrics.detectionmap — Method

detectionmap([method=snr], data, fwhm; fill=0)

Parallel implementation of arbitrary detection mapping applied to each pixel in the input image. Any invalid values will be set to fill.

The following methods are provided in the Metrics module:

snr - signal-to-noise ratio (S/N) using student-t statistics to account for small sample penalty.

significance - Gaussian signifance using student-t statistics to account for samll sample penalty.

Tip

This code is automatically multi-threaded, so be sure to set JULIA_NUM_THREADS before loading your runtime to take advantage of it!

ADI.Metrics.significance — Method

significance(data, position, fwhm)

Calculates the Gaussian significance from the signal-to-noise ratio (SNR, S/N) for a test point at position using apertures of diameter fwhm in a residual frame.

The Gaussian signifiance is calculated from converting the SNR confidence limit from a student t distribution to a Gaussian via

$\text{sig}(\text{SNR}) = \Phi^{-1}\left[\int_0^\text{SNR}{t_\nu(x)dx}\right]$

where the degrees of freedom $\nu$ is given as $2\pi r / \Gamma - 2$ where r is the radial distance of each pixel from the center of the frame.

See Also

snr

ADI.Metrics.snr — Method

snr(data, position, fwhm)

Calculate the signal to noise ratio (SNR, S/N) for a test point at position using apertures of diameter fwhm in a residual frame.

Uses the method of Mawet et al. 2014 which includes penalties for small sample statistics. These are encoded by using a student's t-test for calculating the SNR.

Note

SNR is not equivalent to significance, use significance instead

ADI.Metrics.contrast_curve — Method

contrast_curve

ADI.Metrics.throughput — Method

throughput