Metrics

API/Reference

ADI.Metrics — Module

ADI.Metrics

This module provides code for analyzing the results from ADI in a way that is interpretable statistically. Some of the key functionalities are signal-to-noise, significance, the receiver operating characteristic, and the contrast curve.

ADI.Metrics.significance — Method

significance(snr::AbstractMatrix, fwhm)

Calculates the Gaussian significance from the signal-to-noise ratio (SNR, S/N).

The Gaussian signifiance is calculated from converting the SNR confidence limit from a student t distribution to a Gaussian via

$\text{sig}(\text{SNR}) = \Phi^{-1}\left[\int_0^\text{SNR}{t_\nu(x)dx}\right]$

where the degrees of freedom $\nu$ is given as $2\pi r / \Gamma - 2$ where r is the radial distance of each pixel from the center of the frame.

See Also

snrmap

ADI.Metrics.snr — Method

snr(data, position, fwhm)

Calculate the signal to noise ratio (SNR, S/N) for a test point at position using apertures of diameter fwhm in a residual frame.

Uses the method of Mawet et al. 2014 which includes penalties for small sample statistics. These are encoded by using a student's t-test for calculating the SNR.

Note

SNR is not equivalent to significance, use significance instead

ADI.Metrics.snr_approx! — Method

snr_approx!(data, position, fwhm)

In-place version of snr_approx which modifies data.

ADI.Metrics.snr_approx — Method

snr_approx(data, position, fwhm)

Calculates an approximate signal to noise ratio (SNR, S/N) for a test point at position using apertures of diameter fwhm in a residual frame.

Applies the small sample statistics penalty the same as snr. Data is assumed to have been filtered using a 2D top-hat kernel already (automatically done if called via snrmap)

Note

SNR is not equivalent to significance, use significance instead.

ADI.Metrics.snrmap — Method

snrmap(data, fwhm; method=:exact)

Parallel implementation of signal-to-noise ratio (SNR, S/N) applied to each pixel in the input image.

SNR Methods

:exact - Uses snr to get the exact SNR using student t statistics (small samples penalty)

:approx - Convolves the input by a tophat kernel before using snr_approx! to approximate the student t statistics (small samples penalty).

Tip

This code is automatically multi-threaded, so be sure to set JULIA_NUM_THREADS before loading your runtime to take advantage of it!

ADI.Metrics.contrast_curve — Method

contrast_curve

ADI.Metrics.throughput — Method

throughput